+++ assets/js/omniSearch.js
@@ -1,18 +1,29 @@
 $(document).ready(function ($) {
+  const cacheTimeouts = {
+    projects: parseFloat(omniSearch.settings.projectCacheExpiration),
+    tickets: parseFloat(omniSearch.settings.ticketCacheExpiration),
+  };
   const userId = omniSearch.settings.userId;
   const key = {
     escape: 27,
     period: 190,
     backspace: 8,
   };
+  let isFetching = false; // Is data being fetched
+  let isVisible = false; // Is overlay visible
+
+  // Fetch new data if cache is stale
+  fetchOmnisearchData();
 
   // Append overlay
   $('body').append(`
       <div class="omni-search hidden">
           <select class="js-example-basic-multiple" name="actions[]"></select>
+          <div class="omni-search-panel"></div>
       </div>`);
 
   const omniSelectElement = $('body .omni-search > select');
+  const omniSelectPanelElement = $('body .omni-search > .omni-search-panel');
   $('div.omni-search').on('click', function (e) {
     // Close overlay when clicking outside.
     if ($(e.target).hasClass('omni-search')) {
@@ -45,8 +56,9 @@ $(document).ready(function ($) {
 
   // Init select2, get data, set events.
   function initOmniSearch() {
+    isFetching = false;
+    isVisible = true;
     $('body').addClass('prevent-scroll');
-    omniSelectElement.addClass('loading');
     if ($('.omni-search').hasClass('hidden') === false) {
       return false;
     }
@@ -118,8 +130,7 @@ $(document).ready(function ($) {
               break;
           }
         });
-      getOmnisearchData();
-
+      setOmnisearchData();
       omniSelectElement.on('select2:select', function (e) {
         var selection = e.params.data;
         switch (selection.type) {
@@ -161,6 +172,8 @@ $(document).ready(function ($) {
     omniSelectElement.empty().trigger('change');
     $('body .omni-search').addClass('hidden');
     $('body .select2-container').remove();
+    omniSelectPanelElement.empty();
+    isVisible = false;
   }
 
   // Set up select content based on selected element.
@@ -216,6 +229,9 @@ $(document).ready(function ($) {
 
   // Set data and refresh select2.
   function reinitOmniSearchWithData(data) {
+    if (!isVisible || data.length === 0) {
+      return false;
+    }
     omniSelectElement
       .select2('destroy')
       .empty()
@@ -273,12 +289,8 @@ $(document).ready(function ($) {
     return callApi('leantime.rpc.projects.getAll', {});
   }
 
-  function getUserTickets() {
-    return callApi('leantime.rpc.tickets.getAll', {
-      // searchCriteria: {
-      userId: userId,
-      // },
-    });
+  function getAllTickets() {
+    return callApi('leantime.rpc.tickets.getAll', {});
   }
 
   function callApi(method, params) {
@@ -301,64 +313,138 @@ $(document).ready(function ($) {
     });
   }
 
-  // Get data for initial load.
-  function getOmnisearchData() {
-    var availableTags = [];
+  async function fetchOmnisearchData() {
+    let projectPromise;
+    let ticketPromise;
+    let projectCacheData = getCacheData('projects');
+    isFetching = true;
 
-    let projects = getAllProjects().then((data) => {
-      var projects = data.result;
-      const projectGroup = {
-        id: 'project',
-        text: 'Projects',
-        children: [],
-        index: 1,
-      };
-      projects.forEach((project) => {
-        let option = {
-          id: project.id,
-          text: project.name,
-          type: 'project',
-          client: project.clientName,
+    if (projectCacheData) {
+      projectPromise = Promise.resolve(projectCacheData);
+    } else {
+      projectPromise = getAllProjects().then((data) => {
+        var projects = data.result;
+        const projectGroup = {
+          id: 'project',
+          text: 'Projects',
+          children: [],
+          index: 1,
         };
-        projectGroup.children.push(option);
-      });
-      availableTags.push(projectGroup);
-    });
-
-    let tickets = getUserTickets().then((data) => {
-      var result = data.result;
-      let tickets = result.filter((result) => result.type === 'task');
-      const ticketGroup = {
-        id: 'task',
-        text: 'To-Do´s',
-        children: [],
-        index: 2,
-      };
-      tickets.forEach((ticket) => {
-        let option = {
-          id: ticket.id,
-          text: ticket.headline,
-          type: ticket.type,
-          tags: ticket.tags,
-          sprintName: ticket.sprintName,
-          projectId: ticket.projectId,
-          projectName: ticket.projectName,
-        };
-        ticketGroup.children.push(option);
-      });
-      availableTags.push(ticketGroup);
-    });
-
-    // Sort data by index and return to select2.
-    projects.then(() => {
-      tickets.then(() => {
-        availableTags.sort(function (a, b) {
-          return a.index - b.index;
+        projects.forEach((project) => {
+          let option = {
+            id: project.id,
+            text: project.name,
+            type: 'project',
+            client: project.clientName,
+          };
+          projectGroup.children.push(option);
         });
+        writeToCache('projects', {
+          data: projectGroup,
+          expiration: Date.now(),
+        });
+        return projectGroup;
+      });
+    }
+
+    let ticketCacheData = getCacheData('tickets');
+    if (ticketCacheData) {
+      ticketPromise = Promise.resolve(ticketCacheData);
+    } else {
+      ticketPromise = getAllTickets().then((data) => {
+        var result = data.result;
+        let tickets = result.filter((result) => result.type === 'task');
+        const ticketGroup = {
+          id: 'task',
+          text: 'To-Do´s',
+          children: [],
+          index: 2,
+        };
+        tickets.forEach((ticket) => {
+          let option = {
+            id: ticket.id,
+            text: ticket.headline,
+            type: ticket.type,
+            tags: ticket.tags,
+            sprintName: ticket.sprintName,
+            projectId: ticket.projectId,
+            projectName: ticket.projectName,
+          };
+          ticketGroup.children.push(option);
+        });
+        writeToCache('tickets', {
+          data: ticketGroup,
+          expiration: Date.now(),
+        });
+        return ticketGroup;
+      });
+    }
+
+    const promises = [projectPromise, ticketPromise];
+    const results = await Promise.allSettled(promises);
+    const data = results
+      .filter((result) => result.status === 'fulfilled')
+      .map((result) => result.value)
+      .sort(function (a, b) {
+        // Sort by index.
+        return a.index - b.index;
+      });
+
+    return data;
+  }
+
+  function setOmnisearchData() {
+    omniSelectElement.addClass('loading');
+    if (isFetching) {
+      setTimeout(() => {
+        // If already fetching, recall for cached result.
+        setOmnisearchData();
+      }, 500);
+    } else {
+      fetchOmnisearchData().then((availableTags) => {
+        isFetching = false;
         reinitOmniSearchWithData(availableTags);
         omniSelectElement.removeClass('loading');
+        populateLastUpdated();
       });
+    }
+  }
+
+  function populateLastUpdated() {
+    let projectLastUpdated = readFromCache('projects').expiration;
+    let ticketsLastUpdated = readFromCache('tickets').expiration;
+
+    // Convert ms to minutes
+    let projectsLastUpdatedElement =
+      '<span>Projects: ' +
+      Math.round((Date.now() - projectLastUpdated) / 60000) +
+      ' min ago.</span>';
+    let ticketsLastUpdatedElement =
+      '<span>Tickets: ' +
+      Math.round((Date.now() - ticketsLastUpdated) / 60000) +
+      ' min ago.</span>';
+    omniSelectPanelElement.html(
+      '<div><button id="refreshBtn"><span><i class="fa-solid fa-arrows-rotate"></i>Sync data</span></button></div><div>' +
+        projectsLastUpdatedElement +
+        ticketsLastUpdatedElement +
+        '</div>'
+    );
+
+    $('#refreshBtn').on('click', function (e) {
+      if (isFetching) {
+        return false;
+      }
+      $(this)
+        .children('span')
+        .html('<i class="fa-solid fa-arrows-rotate fa-spin"></i>Syncing data');
+      refreshOmniSearch();
     });
+  }
+
+  function refreshOmniSearch() {
+    removeFromCache('projects');
+    removeFromCache('tickets');
+    setOmnisearchData();
   }
 
   function fuzzySearch(needle, haystack) {
@@ -409,5 +495,32 @@ $(document).ready(function ($) {
     }
 
     return null;
+  }
+
+  function removeFromCache(item) {
+    localStorage.removeItem(item);
+  }
+
+  function writeToCache(item, data) {
+    localStorage.setItem(item, JSON.stringify(data));
+  }
+
+  function readFromCache(item) {
+    return JSON.parse(localStorage.getItem(item)) || null;
+  }
+
+  function getCacheData(item) {
+    const cacheData = readFromCache(item);
+
+    if (!cacheData) {
+      return false;
+    }
+
+    const cacheDataExpiration = cacheData.expiration ?? 0;
+    // Convert minutes to ms
+    const cacheTimeoutMs = cacheTimeouts[item] * 60000;
+    const cacheDataExpired = Date.now() - cacheDataExpiration > cacheTimeoutMs;
+
+    return cacheDataExpired ? false : cacheData.data;
   }
 });