+++ CHANGELOG.md
@@ -7,8 +7,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Used `APP_ROOT` for installing assets.
+
+## [1.2.0] 2024-02-12
+
+### Fixed
+
+- Typo in cache constant.
+
+### Added
+
+- Settings page.
+- Added github actions
+
+## [1.1.0] 2024-02-01
+
+### Added
+
+- Added caching.
+
+### Fixed
+
+- Bug where data would load in after overlay close.
+
+## [1.0.1] 2024-01-31
+
+### Fixed
+
+- Readded searchCriteria api parameter for LT 2.4.8.
+
+## [1.0.0] 2024-01-30
+
 ### Added
 
 - Added OmniSearch plugin.
 - Added fuzzy search.
 - Install, uninstall script.
+
+[1.2.0]: https://github.com/ITK-Leantime/leantime-omnisearch/compare/1.1.0...1.2.0
+[1.1.0]: https://github.com/ITK-Leantime/leantime-omnisearch/compare/1.0.1...1.1.0
+[1.0.1]: https://github.com/ITK-Leantime/leantime-omnisearch/compare/1.0.0...1.0.1
+[1.0.0]: https://github.com/ITK-Leantime/leantime-omnisearch/releases/tag/1.0.0